+++ artemis/src/main/java/com/artemis/EntityTransmuter.java
@@ -121,14 +121,14 @@ public final class EntityTransmuter {
 		private Bag<ComponentType> additions;
 		private Bag<ComponentType> removals;
 		private ComponentManager cm;
-		public final int compositionId;
+		public final short compositionId;
 
 		public TransmuteOperation(ComponentManager cm,
 		                          int compositionId,
 		                          Bag<ComponentType> additions,
 		                          Bag<ComponentType> removals) {
 			this.cm = cm;
-			this.compositionId = compositionId;
+			this.compositionId = (short) compositionId;
 			this.additions = additions;
 			this.removals = removals;
 		}
